api/requirements.txt
alabaster==0.7.10
argh==0.26.2
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
autopep8-autoline==0.1.0
Babel==2.5.3
boto3==1.14.46
botocore==1.17.46
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.5.0
Django==3.0.8
django-braces==1.13.0
django-multiselectfield==0.1.8
django-nose==1.4.6
django-oauth-toolkit==1.1.2
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
docutils==0.14
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
flake8-annotations==2.0.1
flake8-colors==0.1.6
flake8-django==0.0.4
flake8-docstrings==1.5.0
flake8-todo==0.7
forex-python==1.0.0
gunicorn==20.0.4
httpie==0.9.9
idna==2.6
imagesize==1.0.0
inflection==0.5.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10
jmespath==0.10.0
lazy-object-proxy==1.4.3
livereload==2.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.19
mysqlclient==2.0.1
nose==1.3.7
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pathtools==0.1.2
pep8==1.7.1
port-for==0.4
protobuf==3.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
Pyliner==1.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.1
style==1.1.0
tornado==6.0.4
typed-ast==1.4.1
update==0.0.1
uritemplate==3.0.1
urllib3==1.25.8
watchtower==0.8.0
wrapt==1.11.2
